One last refactor, tidy the builds up again
