+++ dlls/xapofx1_1/Makefile.in
@@ -5,6 +5,7 @@ PARENTSRC = ../xaudio2_7
 EXTRALIBS = $(SDL2_LIBS)
 
 C_SRCS = \
+	allocator.c \
 	xapofx.c \
 	xapo.c \
 	FAudio/FAPOBase.c \
@@ -12,8 +13,6 @@ C_SRCS = \
 	FAudio/FAPOFX_eq.c \
 	FAudio/FAPOFX_echo.c \
 	FAudio/FAPOFX_masteringlimiter.c \
-	FAudio/FAPOFX_reverb.c \
-	FAudio/FAudioFX_reverb.c \
-	FAudio/FAudioFX_volumemeter.c
+	FAudio/FAPOFX_reverb.c
 
 RC_SRCS = version.rc
