+++ dlls/xapofx1_2/Makefile.in
@@ -5,6 +5,7 @@ PARENTSRC = ../xaudio2_7
 EXTRALIBS = $(SDL2_LIBS)
 
 C_SRCS = \
+	allocator.c \
 	xapofx.c \
 	xapo.c \
 	FAudio/FAPOBase.c \
@@ -12,6 +13,4 @@ C_SRCS = \
 	FAudio/FAPOFX_eq.c \
 	FAudio/FAPOFX_echo.c \
 	FAudio/FAPOFX_masteringlimiter.c \
-	FAudio/FAPOFX_reverb.c \
-	FAudio/FAudioFX_reverb.c \
-	FAudio/FAudioFX_volumemeter.c
+	FAudio/FAPOFX_reverb.c
