+++ dlls/xaudio2_0/Makefile.in
@@ -5,16 +5,18 @@ PARENTSRC = ../xaudio2_7
 EXTRALIBS = $(SDL2_LIBS) $(LIBAVCODEC_LIBS) $(LIBAVUTIL_LIBS)
 
 C_SRCS = \
+	allocator.c \
 	compat.c \
 	xaudio_dll.c \
 	xaudio_engine.c \
+	xaudio_voice.c \
 	xaudio_voice_source.c \
 	xaudio_voice_submix.c \
 	xaudio_voice_master.c \
-	xaudio_voice_wrap.c \
+	xaudio_voice_wrapeffect.c \
+	xaudio_voice_wrapsend.c \
 	xaudio_callbacks.c \
 	xapo.c \
-	xapo_wrap.c \
 	FAudio/FAPOBase.c \
 	FAudio/FAudio.c \
 	FAudio/FAudio_ffmpeg.c \
