+++ dlls/xaudio2_7/xapofx.c
@@ -1,5 +1,6 @@
 /*
  * Copyright (c) 2015 Andrew Eikum for CodeWeavers
+ * Copyright (c) 2018 Ethan Lee for CodeWeavers
  *
  * This library is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Lesser General Public
@@ -28,7 +29,6 @@
 #endif /* XAPOFX1_VER */
 #include "xaudio_private.h"
 #include "xapofx.h"
-#include "xaudio2fx.h"
 
 #include "wine/debug.h"
 #include "wine/heap.h"
