+++ dlls/xaudio2_7/xaudio_voice_wrapeffect.c
