+++ dlls/xaudio2_8/Makefile.in
@@ -5,18 +5,20 @@ PARENTSRC = ../xaudio2_7
 EXTRALIBS = $(SDL2_LIBS) $(LIBAVCODEC_LIBS) $(LIBAVUTIL_LIBS)
 
 C_SRCS = \
+	allocator.c \
 	compat.c \
 	xapofx.c \
 	x3daudio.c \
 	xaudio_dll.c \
 	xaudio_engine.c \
+	xaudio_voice.c \
 	xaudio_voice_source.c \
 	xaudio_voice_submix.c \
 	xaudio_voice_master.c \
-	xaudio_voice_wrap.c \
+	xaudio_voice_wrapeffect.c \
+	xaudio_voice_wrapsend.c \
 	xaudio_callbacks.c \
 	xapo.c \
-	xapo_wrap.c \
 	FAudio/F3DAudio.c \
 	FAudio/FAPOBase.c \
 	FAudio/FAPOFX.c \
